Update build scripts to support VLC 4.0

+++ .travis/script.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 set -euxo pipefail
 
+echo_readme() {
+  echo "This plugin was built for VLC 4.0, a nightly, i.e. in-development version of VLC." \
+       "The development version of VLC often breaks ABI, making the plugin built for it non-functional in newer VLC builds." \
+       "Meaning that a plugin targeting a VLC 4.0 build done today might not work with tomorrow's build of VLC 4.0." \
+       "Thus it's recommended that you use the specific VLC 4.0 build this plugin was made for: $1" | fold -sw 80
+}
+
 if [ "$TARGET" = "Linux" ]; then
   # start a docker container of a specific version of Debian
   # different Debian versions have different VLC versions packaged: 2.1, 2.2 and 3.0
@@ -35,7 +42,7 @@ if [ "$TARGET" = "Linux" ]; then
   RUN make -C /repo install
 elif [ "$TARGET" = "Windows" ]; then
   sudo apt-get update
-  sudo apt-get install -y tree
+  sudo apt-get install -y tree zip
   # use our own docker image for testing building of Windows binaries
   # makes sure that our docker image is not broken too
   sudo docker build -t vlc-pause-click-plugin-windows-build docker
@@ -44,30 +51,66 @@ elif [ "$TARGET" = "Windows" ]; then
   tree build -s --si --du
   # there should be exactly two dlls built, one for 32-bit VLC and another for 64-bit
   [ $(find ./build -name "*.dll" | wc -l) = "2" ] || false
+  cd build
+  echo_readme "$(cat $VLC_VERSION/32/VLC_DOWNLOAD_URL.txt)" > "$VLC_VERSION/32/README.txt"
+  echo_readme "$(cat $VLC_VERSION/64/VLC_DOWNLOAD_URL.txt)" > "$VLC_VERSION/64/README.txt"
+  ./zip-it.sh
+  cd ..
 elif [ "$TARGET" = "macOS" ]; then
   brew update
   brew install p7zip
   mkdir -p tmp/tmp
   cd tmp/tmp
-  wget https://download.videolan.org/pub/videolan/vlc/$VLC_VERSION.0/macosx/vlc-$VLC_VERSION.0.dmg
-  # extracting contents of a .dmg file using 7-zip is more of a hack, and while it works 7z exits with an error code we want to supress
-  7z x "*.dmg" "*/VLC.app/Contents/MacOS" || true
-  mv */VLC.app/Contents/MacOS/lib ..
-  mv */VLC.app/Contents/MacOS/include ..
-  rm -rf ./*
-  cd ..
-  rm -rf ./tmp
-  # in contrast to Windows VLC including a very nice and complete sdk directory,
-  # macOS VLC doesn't package .pc files, plugin headers and in 2.1.0 libvlc_version.h is not generated off libvlc_version.h.in.
-  # it just packags libvlc.dylib, libvlccore.dylib and libvlc headers.
-  # honestly, wtf VLC team, get your game together.
-  # we grab the missing pieces from the Windows VLC, there shouldn't be anything platform-specific in those so it should work
-  wget https://download.videolan.org/pub/videolan/vlc/$VLC_VERSION.0/win32/vlc-$VLC_VERSION.0-win32.7z -O vlc-$VLC_VERSION.0-win32.7z
-  7z x "*.7z" -o* "*/sdk"
-  mv vlc-$VLC_VERSION.0-win32/*/sdk/include/vlc/libvlc_version.h include/vlc/
-  mv vlc-$VLC_VERSION.0-win32/*/sdk/include/vlc/plugins include/vlc/plugins
-  mv vlc-$VLC_VERSION.0-win32/*/sdk/lib/pkgconfig lib/pkgconfig
-  rm -rf ./vlc-$VLC_VERSION.0-win32
+  if [ "$VLC_VERSION" = "4.0" ]; then
+    brew install lynx
+    LATEST_MACOS_DIR_URL=$(lynx -listonly -nonumbers -dump "https://nightlies.videolan.org/build/macosx-intel/" | grep "vlc-$VLC_VERSION.0-" | LC_COLLATE=C sort --stable --ignore-case | tail -n1)
+    echo "$LATEST_MACOS_DIR_URL"
+    LATEST_MACOS_FILE_URL=$(lynx -listonly -nonumbers -dump "$LATEST_MACOS_DIR_URL" | grep ".dmg$" | tail -n1)
+    echo "$LATEST_MACOS_FILE_URL"
+    if [ "$(echo "$LATEST_MACOS_FILE_URL" | wc -l | awk '{print $1}')" != "1" ]; then
+      echo "Error: Weren't able to find the .dmg file for VLC $VLC_VERSION"
+      exit 1
+    fi
+    wget "$LATEST_MACOS_FILE_URL"
+    7z x "*.dmg" "*/VLC.app/Contents/Frameworks" || true
+    mkdir ../lib
+    mv */VLC.app/Contents/Frameworks/*.dylib ../lib
+    cd ..
+    rm -rf ./tmp
+    LATEST_WIN64_DIR_URL=$(lynx -listonly -nonumbers -dump "https://nightlies.videolan.org/build/win64/" | grep "vlc-$VLC_VERSION.0-" | LC_COLLATE=C sort --stable --ignore-case | tail -n1)
+    echo "$LATEST_WIN64_DIR_URL"
+    LATEST_WIN64_FILE_URL=$(lynx -listonly -nonumbers -dump "$LATEST_WIN64_DIR_URL" | grep ".7z$" | grep -v "debug" | tail -n1)
+    echo "$LATEST_WIN64_FILE_URL"
+    if [ "$(echo "$LATEST_WIN64_FILE_URL" | wc -l | awk '{print $1}')" != "1" ]; then
+      echo "Error: Weren't able to find the .7z file for VLC $VLC_VERSION"
+      exit 1
+    fi
+    wget "$LATEST_WIN64_FILE_URL" -O vlc-$VLC_VERSION.0-win64.7z
+    7z x "*.7z" -o* "*/sdk"
+    mkdir -p include
+    mv vlc-$VLC_VERSION.0-win64/*/sdk/include/vlc include
+    mv vlc-$VLC_VERSION.0-win64/*/sdk/lib/pkgconfig lib
+    rm -rf ./vlc-$VLC_VERSION.0-win64*
+  else
+    wget https://download.videolan.org/pub/videolan/vlc/$VLC_VERSION.0/macosx/vlc-$VLC_VERSION.0.dmg
+    # extracting contents of a .dmg file using 7-zip is more of a hack, and while it works 7z exits with an error code we want to supress
+    7z x "*.dmg" "*/VLC.app/Contents/MacOS" || true
+    mv */VLC.app/Contents/MacOS/lib ..
+    mv */VLC.app/Contents/MacOS/include ..
+    cd ..
+    rm -rf ./tmp
+    # in contrast to Windows VLC including a very nice and complete sdk directory,
+    # macOS VLC doesn't package .pc files, plugin headers and in 2.1.0 libvlc_version.h is not generated off libvlc_version.h.in.
+    # it just packags libvlc.dylib, libvlccore.dylib and libvlc headers.
+    # honestly, wtf VLC team, get your game together.
+    # we grab the missing pieces from the Windows VLC, there shouldn't be anything platform-specific in those so it should work
+    wget https://download.videolan.org/pub/videolan/vlc/$VLC_VERSION.0/win64/vlc-$VLC_VERSION.0-win64.7z -O vlc-$VLC_VERSION.0-win64.7z
+    7z x "*.7z" -o* "*/sdk"
+    mv vlc-$VLC_VERSION.0-win64/*/sdk/include/vlc/libvlc_version.h include/vlc/
+    mv vlc-$VLC_VERSION.0-win64/*/sdk/include/vlc/plugins include/vlc
+    mv vlc-$VLC_VERSION.0-win64/*/sdk/lib/pkgconfig lib
+    rm -rf ./vlc-$VLC_VERSION.0-win64*
+  fi
   # fix paths in .pc files
   sed -i "" "s|^prefix=.*|prefix=$PWD|g" lib/pkgconfig/*.pc
   export PKG_CONFIG_PATH="${PWD}/lib/pkgconfig"
@@ -76,5 +119,11 @@ elif [ "$TARGET" = "macOS" ]; then
   ln -sf libvlccore.*.dylib libvlccore.dylib
   cd ../..
   make OS=macOS
-  zip -j vlc-$VLC_VERSION-macosx.zip libpause_click_plugin.dylib
+  if [ "$VLC_VERSION" = "4.0" ]; then
+    echo "$LATEST_MACOS_DIR_URL" > VLC_DOWNLOAD_URL.txt
+    echo_readme "$(cat VLC_DOWNLOAD_URL.txt)" > "README.txt"
+    zip -j vlc-$VLC_VERSION-macosx.zip libpause_click_plugin.dylib VLC_DOWNLOAD_URL.txt README.txt
+  else
+    zip -j vlc-$VLC_VERSION-macosx.zip libpause_click_plugin.dylib
+  fi
 fi
